### 2025-11-03 #9345 @editorialbot
[PRE REVIEW]: FluxOptics.jl: A Differentiable Wave Optics Framework for Inverse Design in Julia
**Submitting author:** @anscoil (<a href="http://orcid.org/0000-0002-4460-4151">Nicolas BARRE</a>)
**Repository:** https://github.com/anscoil/FluxOptics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
