### 2025-11-30 #9345 @abhishektiwari
@anscoil I will be helping to edit this paper.

Can you bring the word count for `paper.md` to 1000 words or less?

Please also add contribution guideline and to mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/).
