### 2025-11-30 2f495d7
Shortening JOSS article
