+++ paper/paper.md
@@ -21,13 +21,13 @@ bibliography: paper.bib
 
 # Summary
 
-FluxOptics.jl is a Julia package for simulating optical field propagation with full support for automatic differentiation. It enables gradient-based inverse design of optical components, which consists in determining the structure of an optical element (lens, diffraction grating, phase mask) that produces a desired light pattern or functionality. The package implements scalar wave propagation methods that are computationally efficient alternatives to finite-difference time-domain (FDTD) simulations, making it particularly suited for designing optical elements compatible with additive manufacturing techniques such as direct laser writing and two-photon polymerization, as well as low-cost characterization methods like intensity-only diffraction tomography.
+FluxOptics.jl is a Julia package for simulating optical field propagation with full support for automatic differentiation. It enables gradient-based inverse design of optical components, which consists in determining the structure of an optical element (lens, diffraction grating, phase mask) that produces a desired light pattern or functionality. The package implements scalar wave propagation methods that are computationally efficient alternatives to finite-difference time-domain (FDTD) simulations, particularly suited for applications in additive manufacturing and low-cost optical characterization.
 
 FluxOptics.jl provides multiple propagation algorithms for free-space and graded-index media, a composable architecture for building complex optical systems, GPU acceleration, and optimization tools including proximal operators for constrained inverse design. The architecture supports current scalar field applications and is designed to extend to vector field propagation for polarization-dependent components and dielectric metasurfaces.
 
 # Statement of Need
 
-Inverse design of optical components has become increasingly important with the rise of freeform optics [@Schmidt2020; @Barre2025], diffractive optical elements (DOEs) [@Dinc2020], and metasurfaces [@Molesky2018; @Peurifoy2018]. Traditional optimization approaches include gradient-free methods (evolutionary algorithms, Bayesian optimization, stochastic search) which can be effective for low-dimensional problems with up to hundreds of parameters. However, these methods become intractable when the design space scales to thousands or millions of parameters, as is typical for spatially-resolved optical elements defined on computational grids. Gradient-based optimization using automatic differentiation has emerged as the solution for such high-dimensional design spaces [@Hughes2018; @Minkov2020], enabling efficient convergence by exploiting gradient information at computational cost comparable to a single forward simulation.
+Inverse design of optical components has become increasingly important with the rise of freeform optics [@Schmidt2020; @Barre2025], diffractive optical elements (DOEs) [@Dinc2020], and metasurfaces [@Molesky2018; @Peurifoy2018]. Traditional optimization approaches include gradient-free methods (evolutionary algorithms, Bayesian optimization, stochastic search) which can be effective for low-dimensional problems with up to hundreds of parameters. However, these methods become intractable for spatially-resolved optical elements with thousands or millions of parameters. Gradient-based optimization using automatic differentiation [@Hughes2018; @Minkov2020] enables efficient convergence at computational cost comparable to a single forward simulation.
 
 However, existing tools face several limitations. Full-wave electromagnetic solvers like FDTD provide high accuracy but are computationally prohibitive for optimization, often requiring hours per forward simulation and limited to 2D or small 3D domains [@Oskooi2010]. Python packages like TorchOptics [@TorchOptics] provide differentiable scalar wave propagation but suffer from performance bottlenecks.
 
@@ -41,39 +41,25 @@ Fourth, the package implements efficient propagation methods through kernel cach
 
[…]
