+++ README.md
@@ -14,12 +14,20 @@ FluxOptics.jl enables gradient-based optimization of optical systems through ful
 
 ## ✨ Key Features
 
-- 🌊 **Full wave propagation**: Angular Spectrum, Rayleigh-Sommerfeld, Beam Propagation Method
-- 🎯 **Inverse design**: Optimize phase masks, DOEs, and refractive index profiles end-to-end
+- 🌊 **Wave propagation**: Angular Spectrum, Rayleigh-Sommerfeld, Beam Propagation Method
+- 🎯 **Inverse design**: End-to-end optimization of phase masks, DOEs, and refractive index profiles
 - 🔧 **Proximal optimization**: FISTA, TV regularization, ISTA sparsity, custom constraints
-- 📊 **Multi-wavelength**: Simultaneous chromatic optimization and polychromatic beam shaping
-- 🚀 **GPU accelerated**: Seamless CUDA support for large-scale problems
-- 🧩 **Composable**: Intuitive piping syntax for complex optical systems
+- 📊 **Multi-wavelength & multimode**: Polychromatic propagation and mode coupling
+- 🚀 **GPU accelerated**: Seamless CUDA support
+- 🧩 **Composable architecture**: Intuitive piping syntax for complex optical systems
+
+## Applications
+
+FluxOptics.jl is designed for:
+- **Inverse design**: Diffractive optical elements, metasurfaces, beam shaping
+- **Optical characterization**: Phase retrieval, tomographic reconstruction, waveguide/fiber analysis
+- **Laser physics**: Cavity eigenmode analysis, mode selection, resonator design
+- **Photonics**: Fiber coupling, multimode decomposition, GRIN media simulation
 
 ## 📦 Installation
 
@@ -28,38 +36,34 @@ using Pkg
 Pkg.add("FluxOptics")
 ```
 
-## 🚀 Quick Start: Optimize a Beam Splitter
+## 🚀 Quick Start
 
-Design diffractive optical elements that split a Gaussian beam into two equal output beams:
+Design diffractive optical elements that split a Gaussian beam into two equal outputs:
 
 ```julia
 using FluxOptics, Zygote
-using CUDA  # Comment if you don't have CUDA
 
-# Define target: two separated Gaussians
+# Setup: 512×512 grid, 1 μm pitch, 1.064 μm wavelength
 ns = (512, 512)
 ds = (1.0, 1.0)
 λ = 1.064
 x, y = spatial_vectors(ns, ds)
+
+# Source and target fields
 u0 = ScalarField(Gaussian(30.0)(x, y), ds, λ)
-target_mode = ScalarField(Gaussian(30.0)(x, y, Shift2D(-60, 0)), ds, λ) +
-              ScalarField(Gaussian(30.0)(x, y, Shift2D(60, 0)), ds, λ)
+target = ScalarField(Gaussian(30.0)(x, y, Shift2D(-60, 0)), ds, λ) +
+         ScalarField(Gaussian(30.0)(x, y, Shift2D(60, 0)), ds, λ)
 normalize_power!(u0)
-normalize_power!(target_mode)
+normalize_power!(target)
 
-# Comment if you don't have CUDA
-u0 = cu(u0)
-target_mode = cu(target_mode)
-
-# Optical system: source → trainable DOEs → propagation
+# Optical system: source → DOE → propagation → DOE → propagation
 doe1 = Phase(u0, zeros(size(u0)); trainable=true, buffered=true)
 doe2 = Phase(u0, zeros(size(u0)); trainable=true, buffered=true)
-prop1 = RSProp(u0, 1500.0)
-prop2 = RSProp(u0, 2000.0)
[…]
