### 2025-11-30 4a352a6
Simplifying README
