### 2025-12-15 230086b
Correction of code example in README
