### 2025-12-16 #9345 @Windier
Hi @abhishektiwari. I'm happy to review it. I'm a third year PhD student working on applications of Optical Inverse Design (topic of the library) and have used Julia for ~2 years in my research.
