### 2025-12-29 #9345 @abhishektiwari
👋 @MikeHughesKent @KBodolai @HarrisonKramer, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "FluxOptics.jl: A Differentiable Wave Optics Framework for Inverse Design in Julia".

Thanks
Abhi
