### 2026-01-05 #9345 @aquilesC
Perhaps @mdmaas can be introduced to JOSS?
