### 2026-02-03 #9734 @mdmaas
> Hi [@Windier](https://github.com/Windier), [@mdmaas](https://github.com/mdmaas) checking how you progressing with review. Feel free to create a new issue in the [target repository](https://github.com/anscoil/FluxOptics.jl) and link to those issues. Let me know if any questions.

Hi @abhishektiwari! I'm just back from vacations, and it is also my first review here on JOSS, so I still need to get familiar with the process. I know FluxOptics.jl reasonably well, so just give me a few weeks to accomodate.
