### 2026-02-08 #9734 @Windier
FluxOptics.jl is an optics inverse design package for differentiable optics applications. The software is a very welcome contribution to the inverse design optics community. It is well written, well documented, and the examples cover a good range of applications.

The software paper is also well written and addresses all points of the checklist.

The API is clean and makes good use of Julia’s syntax.

I have verified the FluxOptics performance claims. The example [here](https://anscoil.github.io/FluxOptics.jl/stable/api/#Typical-Workflow:-Beam-Splitter) runs in 0.45 seconds (1.28 M allocations: 64.982 MiB, 4.10% GC time) on an RTX 3070 GPU, which is on par with what is reported there given the GPU difference. The TorchOptics example [here](https://torchoptics.readthedocs.io/en/stable/auto_examples/training_diffractive_splitter.html) ran on 6.4s.

I recommend publication of FluxOptics.jl in JOSS. I only have minor comments about the submission.

Comments:

- I see that you correctly have the references on [paper.bib](https://github.com/anscoil/FluxOptics.jl/blob/main/paper/paper.bib), but is the reference section supposed to be empty in [paper.md](https://github.com/anscoil/FluxOptics.jl/blob/main/paper/paper.md)? (Edit on 09/02: Ignore this one, the references are there in the generated PDF).

- One of the classic inverse design examples is multimode reshaping of Hermite–Gaussian beams, as seen in [1] and [2]. This would be a good addition to the current examples, as it has a level of complexity comparable to what is already included and is of interest to the Multi-Plane Light Converter community, which represents a significant portion of the optical inverse design field. This addition is not conditional on my acceptance and is up to the author to decide.

- CairoMakie is used in all examples but isn’t listed as a FluxOptics dependency. I know it's obvious, but It might be worth adding a comment about optional dependencies, so users know they need to install CairoMakie to run the plotting examples.

- Is a minimum Julia version requirement stated somewhere? Is it 1.11.x? I tested the code on Julia 1.11.7, which is specified in Project.toml.

- I believe that on a fresh Julia installation with Pkg.add("FluxOptics"), the code at
https://anscoil.github.io/FluxOptics.jl/stable/api/#1.-Setup:-Input-and-Target-modes
will not work because Zygote and Optimisers need to be added separately.

References
[1] Laguerre–Gaussian mode sorter. Nature Communications 10, 1865 (2019).
[2] Inverse design of gradient-index volume multimode converters. Optics Express, Vol. 30, Issue 7, pp. 10573–10587.
