### 2026-02-09 #9734 @mdmaas
FluxOptics.jl is a well-designed, high-performance differentiable wave optics framework for inverse design in Julia. The software makes a substantial contribution by providing a composable architecture with significant performance advantages over existing tools. The paper is well-written and the software demonstrates careful engineering with comprehensive documentation.

I recommend publication in JOSS with a few observations:

#### 1 Documentation dependencies

The software passes all the items in the checklist with a minor exception:

The Quick Start example in the README and the API documentation example (https://anscoil.github.io/FluxOptics.jl/stable/api/#1.-Setup:-Input-and-Target-modes) both require `Zygote`, `Optimisers`, and `CairoMakie`, but these are not listed as dependencies in the main Project.toml (only in docs/Project.toml). A fresh user running `Pkg.add("FluxOptics")` will encounter import errors when trying the first example.

Recommendation: Add a brief note in the README installation section, such as:
```julia
# For running optimization examples, also install:
Pkg.add(["Zygote", "Optimisers", "CairoMakie"])
```
Alternatively, add a note like: "Note: Examples require Zygote.jl for automatic differentiation, Optimisers.jl for optimization algorithms, and CairoMakie.jl for visualization."

#### 2. Missing Comparison Context
The paper compares FluxOptics.jl with TorchOptics (Python) and mentions FDTD solvers, but doesn't discuss other Julia optics packages if any exist, or explicitly state if FluxOptics is the first/only differentiable optics package in Julia.

Recommendation: Add a brief statement clarifying FluxOptics.jl's position in the Julia ecosystem. If it's the first differentiable optics package in Julia, state this explicitly to help readers understand its significance.

#### 3. Tutorial Execution Times (Optional)

None of the five tutorials specify expected execution times, which would help users gauge computational requirements and verify successful execution.

Recommendation: Consider adding approximate run times to tutorial introductions (e.g., "This tutorial takes approximately 2-3 minutes on a modern CPU").
