+++ Project.toml
@@ -1,7 +1,7 @@
 name = "FluxOptics"
 uuid = "5d1b044d-dcff-404c-be27-e1944ce84415"
-authors = ["Nicolas Barré"]
 version = "0.1.0"
+authors = ["Nicolas Barré"]
 
 [deps]
 AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
@@ -27,28 +27,28 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-AbstractFFTs = "1.5.0"
-CUDA = "5.8.2"
-ChainRulesCore = "1.25.2"
-CoordinateTransformations = "0.6.4"
-EllipsisNotation = "1.8.0"
-FFTW = "1.9.0"
-FINUFFT = "3.4.2"
-Functors = "0.5.2"
-JuliaFormatter = "2.1.6"
-LRUCache = "1.6.2"
-LaTeXStrings = "1.4.0"
-LazyArrays = "2.6.2"
-LinearAlgebra = "1.11.0"
-Makie = "0.24.4"
-Optimisers = "0.4.6"
-Polynomials = "4.1.0"
-Requires = "1.3.1"
-Rotations = "1.7.1"
-StaticArrays = "1.9.14"
-Statistics = "1.11.1"
-Zygote = "0.7.10"
 julia = "1.11"
+AbstractFFTs = "1.5"
+CUDA = "5.8"
+ChainRulesCore = "1.25"
+CoordinateTransformations = "0.6"
+EllipsisNotation = "1.8"
+FFTW = "1.9"
+FINUFFT = "3.4"
+Functors = "0.5"
+JuliaFormatter = "2.1"
+LRUCache = "1.6"
+LaTeXStrings = "1.4"
+LazyArrays = "2.6"
+LinearAlgebra = "1.11"
+Makie = "0.24"
+Optimisers = "0.4"
+Polynomials = "4.1"
+Requires = "1.3"
+Rotations = "1.7"
+StaticArrays = "1.9"
+Statistics = "1.11"
+Zygote = "0.7"
 
 [extras]
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
