### 2026-02-11 aa18881
Correcting README and updating CI with julia 1.12
