+++ .github/workflows/CI.yml
@@ -23,7 +23,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - ['1.11', '1.12']
+          - '1.11'
+	  - '1.12'
         os:
           - ubuntu-latest
         arch:
@@ -48,7 +49,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: julia-actions/setup-julia@v2
         with:
-          version: ['1.11', '1.12']
+          version: '1.12'
       - uses: julia-actions/cache@v2
       - name: Configure doc environment
         shell: julia --project=docs --color=yes {0}
