+++ .github/workflows/CI.yml
@@ -24,7 +24,7 @@ jobs:
       matrix:
         version:
           - '1.11'
-	  - '1.12'
+          - '1.12'
         os:
           - ubuntu-latest
         arch:
