+++ Project.toml
@@ -51,7 +51,6 @@ Statistics = "1.11"
 Zygote = "0.7"
 
 [extras]
-Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
