### 2026-02-12 deefee9
Adding tutorial on mode sorting
