+++ docs/literate/06_multimode_HG_sorter.jl
@@ -1,0 +1,256 @@
+# # Hermite-Gaussian multimode sorter
+#
+# This tutorial demonstrates the inverse design of a multi-plane diffractive optical system
+# that transforms spatially separated Gaussian beams into Hermite-Gaussian modes. This
+# canonical mode sorting problem has applications in mode-division multiplexing
+# for telecommunications and multimode fiber imaging.
+#
+# The approach presented here builds on the work of Carpenter and Fontaine, who demonstrated
+# at CLEO/Europe in **June 2017** [1] an efficient mapping from triangular Gaussian beam arrays
+# to Laguerre-Gaussian modes (45 modes) using a remarkably low number of phase planes. This work
+# was significantly scaled up in their Nature Communications paper [2], demonstrating high-fidelity
+# mode sorting of **210 modes** with only a modest increase in the number of planes. This remarkable
+# achievement showcases the efficiency of this approach.
+#
+# At that time, gradient descent was not yet widespread in the optics
+# community outside of nanophotonics. The original designs relied on error-reduction
+# algorithms similar to Gerchberg-Saxton, popular since the 1990s. With today's automatic
+# differentiation tools, we can revisit these problems using modern gradient-based optimization.
+# This approach has been explored in recent works [3,4] and enables rapid design iteration
+# as demonstrated in this tutorial.
+#
+# ## Technical approach
+#
+# The optical system consists of multiple phase masks (Diffractive Optical Elements, or DOEs)
+# separated by free-space propagation. This configuration is sometimes referred to as multi-plane
+# light conversion (MPLC) or Diffractive Optical Neural Networks (DONNs) in different contexts.
+# In this tutorial, we optimize 8 cascaded phase masks using algorithmic differentiation,
+# a mathematical optimization approach applicable to any diffractive optical system.
+#
+# **References:**
+#
+# [1]. J. Carpenter and N. Fontaine,
+#    "[Multi-plane light conversion with low plane count](https://www.youtube.com/watch?v=iXjFQj15Xzg),"
+#    CLEO/Europe, **EJ-2** (2017).
+#
+# [2]. N. K. Fontaine, R. Ryf, H. Chen, D. T. Neilson, K. Kim, and J. Carpenter,
+#    "[Laguerre-Gaussian mode sorter](https://doi.org/10.1038/s41467-019-09840-4)",
+#    Nature communications, **10(1)**, 1865 (2019).
+#
+# [3]. N. Barré and A. Jesacher,
+#    "[Inverse design of gradient-index volume multimode converters](https://doi.org/10.1364/OE.450196),"
+#    Optics Express, **30(7)**, 10573-10587 (2022).
+#
+# [4]. H. Kupianskyi, S. A. Horsley and D. B. Phillips,
+#    "[High-dimensional spatial mode sorting and optical circuit design using multi-plane light conversion](https://doi.org/10.1063/5.0128431),"
+#    Apl Photonics, **8(2)** (2023).
+
+using FluxOptics, Zygote, CairoMakie
+using CUDA  #src
+CUDA.allowscalar(false)  #src
+#nb using CUDA  # comment if you don't have CUDA
+prefix = "06_multimode_HG_sorter"  #src
+
+# ## Problem setup
+#
+# We define the computational grid and wavelength for the optical system.
+
+ns = (350, 400)
+ds = (1.0, 1.0)
+λ = 1.55
+
+# ## Input modes: Gaussian beams in triangular layout
+#
[…]
