+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-FluxOptics.jl is a Julia package for simulating optical field propagation with full support for automatic differentiation. It enables gradient-based inverse design of optical components, which consists in determining the structure of an optical element (lens, diffraction grating, phase mask) that produces a desired light pattern or functionality. The package implements scalar wave propagation methods that are computationally efficient alternatives to finite-difference time-domain (FDTD) simulations, particularly suited for applications in additive manufacturing and low-cost optical characterization.
+FluxOptics.jl is a Julia package for simulating optical field propagation with full support for automatic differentiation. It enables gradient-based inverse design of optical components, which consists in determining the structure of an optical element (lens, diffraction grating, phase mask) that produces a desired light pattern or functionality. The package implements scalar wave propagation methods that are computationally efficient alternatives to finite-difference time-domain (FDTD) simulations, particularly suited for applications in holography, additive manufacturing and optical characterization.
 
 FluxOptics.jl provides multiple propagation algorithms for free-space and graded-index media, a composable architecture for building complex optical systems, GPU acceleration, and optimization tools including proximal operators for constrained inverse design. The architecture supports current scalar field applications and is designed to extend to vector field propagation for polarization-dependent components and dielectric metasurfaces.
 
@@ -29,19 +29,20 @@ FluxOptics.jl provides multiple propagation algorithms for free-space and graded
 
 Inverse design of optical components has become increasingly important with the rise of freeform optics [@Schmidt2020; @Barre2025], diffractive optical elements (DOEs) [@Dinc2020], and metasurfaces [@Molesky2018; @Peurifoy2018]. Traditional optimization approaches include gradient-free methods (evolutionary algorithms, Bayesian optimization, stochastic search) which can be effective for low-dimensional problems with up to hundreds of parameters. However, these methods become intractable for spatially-resolved optical elements with thousands or millions of parameters. Gradient-based optimization using automatic differentiation [@Hughes2018; @Minkov2020] enables efficient convergence at computational cost comparable to a single forward simulation.
 
-However, existing tools face several limitations. Full-wave electromagnetic solvers like FDTD provide high accuracy but are computationally prohibitive for optimization, often requiring hours per forward simulation and limited to 2D or small 3D domains [@Oskooi2010]. Python packages like TorchOptics [@TorchOptics] provide differentiable scalar wave propagation but suffer from performance bottlenecks.
+However, existing tools face several limitations. Full-wave electromagnetic solvers like FDTD provide high accuracy but are computationally prohibitive for optimization, often requiring hours per forward simulation and limited to 2D or small 3D domains [@Oskooi2010]. Python packages like TorchOptics [@TorchOptics] provide differentiable scalar wave propagation but suffer from performance bottlenecks. In Julia, WaveOpticsPropagation.jl [@Wechsler:24] offers individual propagation functions, while FluxOptics.jl provides an integrated inverse design framework combining unified propagation components with established optimization methodology (proximal operators, FISTA acceleration) for end-to-end optical system design.
 
-FluxOptics.jl addresses these gaps through several key innovations. First, it provides high-performance differentiable propagation via Zygote.jl [@Innes2019] or Enzyme.jl [@Moses2021]. Implemented in Julia [@Bezanson2017], the package achieves CPU and GPU implementations that significantly outperform existing Python-based tools.
+FluxOptics.jl addresses these gaps through several key innovations. Implemented in Julia [@Bezanson2017] with custom adjoint rules via ChainRulesCore.jl, the package achieves high-performance CPU and GPU implementations that outperform 
[…]
