+++ paper/paper.bib
@@ -58,6 +58,22 @@
   doi={10.1016/j.cpc.2009.11.008}
 }
 
+
+@article{Wechsler:24,
+author = {Felix Wechsler and Carlo Gigli and Jorge Madrid-Wolff and Christophe Moser},
+journal = {Optics. Express},
+keywords = {3D printing; Computed tomography; Liquid crystal displays; Material properties; Ray tracing; Refractive index},
+number = {8},
+pages = {14705--14712},
+publisher = {Optica Publishing Group},
+title = {Wave optical model for tomographic volumetric additive manufacturing},
+volume = {32},
+month = {Apr},
+year = {2024},
+url = {https://opg.optica.org/oe/abstract.cfm?URI=oe-32-8-14705},
+doi = {10.1364/OE.521322},
+}
+
 @misc{TorchOptics,
       title={TorchOptics: An open-source Python library for differentiable Fourier optics simulations},
       author={Matthew J. Filipovich and A. I. Lvovsky},
@@ -201,4 +217,16 @@
   year={2020},
   publisher={De Gruyter},
   doi={10.1515/nanoph-2020-0196}
+}
+
+@article{Fontaine2019,
+  title={Laguerre-Gaussian mode sorter},
+  author={Fontaine, Nicolas K and Ryf, Roland and Chen, Haoshuo and Neilson, David T and Kim, Kwangwoong and Carpenter, Joel},
+  journal={Nature communications},
+  volume={10},
+  number={1},
+  pages={1865},
+  year={2019},
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41467-019-09840-4}
 }
