### 2026-02-12 #9734 @anscoil
Thanks for the positive feedback. I've updated the main branch to address your comments.

## Responses to @Windier

**References in paper.bib**: The compilation process takes care of adding the references properly, so this should be working as expected.

**Tutorial 6 (45-mode sorter)**: I've added a new tutorial based on your references demonstrating a 45-mode conversion system. The tutorial is available in the [development documentation](https://anscoil.github.io/FluxOptics.jl/dev/) (which will become stable when I create a new version after the review process) and is also available in the `notebooks/` folder.

**Visualization dependencies**: I've modified the README.md to clarify that CairoMakie must be installed to use `visualize()` and GLMakie for the interactive version `visualize_slider()`.

**Julia version requirement**: I've added a mention in README.md that Julia version must be ≥ 1.11, which is also specified in the Project.toml as you mentioned. I also added tests for version 1.12 in GitHub CI.

**Manual installation of Zygote and Optimisers**: You're absolutely right. I've modified the README.md to specify that these packages should be installed manually. They are included in the project dependencies, but this is not sufficient due to Julia's intentional design behavior.

## Responses to @mdmaas

**1. Installation dependencies**

This issue was also identified by @Windier, and I've updated the installation section of README.md to provide clear guidance on this.

**2. Positioning in the Julia ecosystem**

I cannot claim that FluxOptics.jl is the first or only differentiable wave optics package in Julia. That would be overstating the case. In Julia, it's relatively straightforward to achieve differentiability by writing pure functions and using Zygote.

However, I believe FluxOptics.jl is currently the only package offering this integrated feature set for optical inverse design. The uniqueness comes from several aspects: the modular architecture that unifies diverse optical components under a common interface, the ability to quickly prototype using the PureComponent pattern while having access to production-level performance through the CustomComponent interface with manual gradient control (using an interface more user-friendly than raw ChainRulesCore.jl), and the extension of Optimisers.jl to handle joint optimization with potentially different optimizers per optical component.

In the Julia community, the only package I found that provides differentiable propagators and is actively maintained is WaveOpticsPropagation.jl. However, its scope is limited to propagation functions. It doesn't provide optical components, system-level abstractions, or field classes. I've modified the JOSS paper to acknowledge this package:

```
In Julia, WaveOpticsPropagation.jl [@Wechsler:24] offers individual propagation functions,
while FluxOptics.jl provides an integrated inverse design framework combining unified
propagation components with established optimization methodology (proximal operators,
FISTA acceleration) for end-to-end optical system design.
```

Given the word count limitations in JOSS articles, I believe what I've mentioned here is adequately captured in the paper. I hope this modification along with my explanation addresses your concern.

**3. Tutorial execution times**

You're right that this is important. FluxOptics.jl is designed to be fast and address real-world problems, so users should see evidence of this. For maintainability reasons, I generate and run my tutorials locally on GPU using Literate.jl rather than through GitHub CI. The script generates markdown files for the documentation (which are not executed and thus don't contain timing), but it also generates Jupyter notebooks that are fully executed and include timing information. I recognize this wasn't obvious from the documentation structure.

I've added an index entry in the tutorials section of the [development documentation](https://anscoil.github.io/FluxOptics.jl/dev/) that highlights the notebook repository and provides timing benchmarks on my GPU:

```
Execution Times (GeForce RTX 4070 Super):

- Tutorials 1-4: 2-6 seconds each
- Tutorial 5: ~42 seconds for the full 105-mode example
- Tutorial 6: ~34 seconds for the 45-mode conversion with 8 planes

CPU execution is possible but significantly slower and not recommended for
iterative development.
```

Please let me know if you need any clarifications or have additional comments.

Best,
Nicolas
