### 2026-02-13 #9734 @Windier
@abhishektiwari, The author has addressed all my comments, and I recommend the publication of FluxOptics.jl in JOSS.
