### 2026-02-13 #9734 @michaellinhq
Hi, I reviewed '[REVIEW]: FluxOptics.jl: A Differentiable Wave Optics Framework for Inverse Design in Julia' and can start immediately. I can deliver a first usable version quickly and iterate based on your feedback. If this scope still matches your priority, I can proceed right away.
