+++ src/optical_components/freespace_propagators/angular_spectrum.jl
@@ -1,27 +1,20 @@
-function as_kernel(fx::T, fy::T, λ::T, n0::Tp, z::Tp, filter::H,
-                   z_pos::Val{true}) where {T <: Real, Tp <: Real, H}
+function as_kernel(fx::T, fy::T, λ::T, n0::Tp, z::Tp,
+                   filter::H) where {T <: Real, Tp <: Real, H}
     fx, fy, λ = Tp(fx), Tp(fy), Tp(λ)/n0
     f² = complex(inv(λ^2))
     v = isnothing(filter) ? Complex{Tp}(1) : Complex{Tp}(filter(fx, fy))
-    Complex{T}(cis(Tp(2)*π*z*sqrt(f² - fx^2 - fy^2)) * v)
+    kernel = z > 0 ? cis(Tp(2)*π*z*sqrt(f² - fx^2 - fy^2)) :
+             conj(cis(-Tp(2)*π*z*sqrt(f² - fx^2 - fy^2)))
+    Complex{T}(kernel * v)
 end
 
-function as_kernel(fx::T, fy::T, λ::T, n0::Tp, z::Tp, filter::H,
-                   z_pos::Val{false}) where {T <: Real, Tp <: Real, H}
-    conj(as_kernel(fx, fy, λ, n0, -z, filter, Val(true)))
-end
-
-function as_kernel(fx::T, λ::T, n0::Tp, z::Tp, filter::H,
-                   z_pos::Val{true}) where {T <: Real, Tp <: Real, H}
+function as_kernel(fx::T, λ::T, n0::Tp, z::Tp, filter::H) where {T <: Real, Tp <: Real, H}
     fx, λ = Tp(fx), Tp(λ)/n0
     f² = complex(inv(λ)^2)
     v = isnothing(filter) ? Complex{Tp}(1) : Complex{Tp}(filter(fx))
-    Complex{T}(cis(Tp(2)*π*z*sqrt(f² - fx^2)) * v)
-end
-
-function as_kernel(fx::T, λ::T, n0::Tp, z::Tp, filter::H,
-                   z_pos::Val{false}) where {T <: Real, Tp <: Real, H}
-    conj(as_kernel(fx, λ, n0, -z, filter, Val(true)))
+    kernel = z > 0 ? cis(Tp(2)*π*z*sqrt(f² - fx^2)) :
+             conj(cis(-Tp(2)*π*z*sqrt(f² - fx^2)))
+    Complex{T}(kernel * v)
 end
 
 function as_paraxial_kernel(fx::T, fy::T, λ::T, θx::T, θy::T, track_tilts::Bool,
@@ -43,36 +36,28 @@ function as_paraxial_kernel(fx::T, λ::T, θx::T, track_tilts::Bool, n0::Tp, z::
 end
 
 function tilted_as_kernel(fx::T, fy::T, λ::T, θx::T, θy::T, track_tilts::Bool,
-                          n0::Tp, z::Tp, filter::H,
-                          z_pos::Val{true}) where {T <: Real, Tp <: Real, H}
+                          n0::Tp, z::Tp, filter::H) where {T <: Real, Tp <: Real, H}
     fx, fy, λ = Tp(fx), Tp(fy), Tp(λ)/n0
     θx, θy = Tp(θx), Tp(θy)
     f² = complex(inv(λ)^2)
     f0x, f0y = sin(θx)/λ, sin(θy)/λ
     offset = track_tilts ? (f0x*fx + f0y*fy)*λ : Tp(0)
[…]
