### 2026-02-14 871f135
Normalization fusion in AS and RS Fourier propagators
