+++ src/optical_components/freespace_propagators/angular_spectrum.jl
@@ -1,42 +1,45 @@
 function as_kernel(fx::T, fy::T, λ::T, n0::Tp, z::Tp,
-                   filter::H) where {T <: Real, Tp <: Real, H}
+                   filter::H, nrm_f::Tp) where {T <: Real, Tp <: Real, H}
     fx, fy, λ = Tp(fx), Tp(fy), Tp(λ)/n0
     f² = complex(inv(λ^2))
     v = isnothing(filter) ? Complex{Tp}(1) : Complex{Tp}(filter(fx, fy))
     kernel = z > 0 ? cis(Tp(2)*π*z*sqrt(f² - fx^2 - fy^2)) :
              conj(cis(-Tp(2)*π*z*sqrt(f² - fx^2 - fy^2)))
-    Complex{T}(kernel * v)
+    Complex{T}(kernel * v * nrm_f)
 end
 
-function as_kernel(fx::T, λ::T, n0::Tp, z::Tp, filter::H) where {T <: Real, Tp <: Real, H}
+function as_kernel(fx::T, λ::T, n0::Tp, z::Tp, filter::H,
+                   nrm_f::Tp) where {T <: Real, Tp <: Real, H}
     fx, λ = Tp(fx), Tp(λ)/n0
     f² = complex(inv(λ)^2)
     v = isnothing(filter) ? Complex{Tp}(1) : Complex{Tp}(filter(fx))
     kernel = z > 0 ? cis(Tp(2)*π*z*sqrt(f² - fx^2)) :
              conj(cis(-Tp(2)*π*z*sqrt(f² - fx^2)))
-    Complex{T}(kernel * v)
+    Complex{T}(kernel * v * nrm_f)
 end
 
 function as_paraxial_kernel(fx::T, fy::T, λ::T, θx::T, θy::T, track_tilts::Bool,
-                            n0::Tp, z::Tp, filter::H) where {T <: Real, Tp <: Real, H}
+                            n0::Tp, z::Tp, filter::H,
+                            nrm_f::Tp) where {T <: Real, Tp <: Real, H}
     fx, fy, λ = Tp(fx), Tp(fy), Tp(λ/n0)
     f0x, f0y = sin(θx)/λ, sin(θy)/λ
     offset = track_tilts ? 2*(f0x*fx + f0y*fy) : Tp(0)
     v = isnothing(filter) ? Complex{Tp}(1) : Complex{Tp}(filter(fx, fy))
-    Complex{T}(cis(-π*λ*z*((fx+f0x)^2 + (fy+f0y)^2) - offset + Tp(2)*π*z/λ) * v)
+    Complex{T}(cis(-π*λ*z*((fx+f0x)^2 + (fy+f0y)^2) - offset + Tp(2)*π*z/λ) * v * nrm_f)
 end
 
 function as_paraxial_kernel(fx::T, λ::T, θx::T, track_tilts::Bool, n0::Tp, z::Tp,
-                            filter::H) where {T <: Real, Tp <: Real, H}
+                            filter::H, nrm_f::Tp) where {T <: Real, Tp <: Real, H}
     fx, λ = Tp(fx), Tp(λ/n0)
     f0x = sin(θx)/λ
     offset = track_tilts ? 2*f0x*fx : Tp(0)
     v = isnothing(filter) ? Complex{Tp}(1) : Complex{Tp}(filter(fx))
-    Complex{T}(cis(-π*λ*z*(fx+f0x)^2 - offset + Tp(2)*π*z/λ) * v)
+    Complex{T}(cis(-π*λ*z*(fx+f0x)^2 - offset + Tp(2)*π*z/λ) * v * nrm_f)
 end
 
[…]
