+++ src/optical_components/freespace_propagators/rayleigh_sommerfeld.jl
@@ -1,29 +1,22 @@
-function rs_kernel(x::T, y::T, λ::T, z::Tp, nrm_f::Tp,
-                   z_pos::Val{true}) where {T <: Real, Tp <: Real}
+function rs_kernel(x::T, y::T, λ::T, z::Tp, nrm_f::Tp) where {T <: Real, Tp <: Real}
     x, y = Tp(x), Tp(y)
     k = Tp(2π/λ)
     r = sqrt(x^2 + y^2 + z^2)
-    Complex{T}(nrm_f*(cis(k*r)/r)*(z/r)*(1/r-im*k))
-end
-
-function rs_kernel(x::T, y::T, λ::T, z::Tp, nrm_f::Tp,
-                   z_pos::Val{false}) where {T <: Real, Tp <: Real}
-    conj(rs_kernel(x, y, λ, -z, nrm_f, Val(true)))
+    kernel = z > 0 ? (cis(k*r)/r)*(z/r)*(1/r-im*k) :
+             conj(cis(k*r)/r)*(-z/r)*(1/r+im*k)
+    Complex{T}(kernel * nrm_f)
 end
 
 function rs_tilted_kernel(x::T, y::T, λ::T, θx::T, θy::T, track_tilts::Bool, z::Tp,
-                          nrm_f::Tp, z_pos::Val{true}) where {T <: Real, Tp <: Real}
+                          nrm_f::Tp) where {T <: Real, Tp <: Real}
     x, y = Tp(x), Tp(y)
     f0x, f0y = sin(θx)/λ, sin(θy)/λ
     k = Tp(2π/λ)
     r = sqrt(x^2 + y^2 + z^2)
     lin_phase = track_tilts ? Complex{Tp}(1) : cis(-Tp(2)*π*(x*f0x+y*f0y))
-    Complex{T}(nrm_f*(cis(k*r)/r)*lin_phase*(z/r)*(1/r-im*k))
-end
-
-function rs_tilted_kernel(x::T, y::T, λ::T, θx::T, θy::T, track_tilts::Bool,
-                          z::Tp, nrm_f::Tp, z_pos::Val{false}) where {T <: Real, Tp <: Real}
-    conj(rs_tilted_kernel(x, y, λ, θx, θy, track_tilts, -z, nrm_f, Val(true)))
+    kernel = z > 0 ? (cis(k*r)/r)*lin_phase*(z/r)*(1/r-im*k) :
+             conj(cis(k*r)/r)*lin_phase*(-z/r)*(1/r+im*k)
+    Complex{T}(kernel * nrm_f)
 end
 
 function rs_valid_distance(nx, ny, dx, dy, λ)
@@ -49,10 +42,11 @@ struct RSKernelProp{M, K, T, Tp} <: AbstractPropagator{M, K, T}
                           double_precision_kernel::Bool
                           = use_cache) where {T, U <: AbstractArray{Complex{T}}, Nd}
         ns = size(u)[1:Nd]
+        ns′ = map(n -> 2*n-1, ns)
         cache_size = use_cache ? prod(size(u)[(Nd + 1):end]) : 0
-        kernel = ConvolutionKernel(u.electric, ns, ds, cache_size)
+        kernel = ConvolutionKernel(u.electric, ns, ds, cache_size; normalize = false)
         Tp = double_precision_kernel ? Float64 : T
-        nrm_f = Tp(prod(ds)/2π)
+        nrm_f = Tp(prod(ds)/2π/prod(ns′))
         new{Static, typeof(kernel), T, Tp}(kernel, track_tilts, Tp(z), nrm_f)
     end
 end
@@ -71,9 +65,9 @@ end
 
[…]
