+++ benchmarks/angular_spectrum/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+FluxOptics = "5d1b044d-dcff-404c-be27-e1944ce84415"
+NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
