### 2026-02-16 6a20f05
Benchmarking Angular Spectrum
