+++ benchmarks/angular_spectrum/scripts/bench_waveopticspropagation.jl
@@ -1,0 +1,145 @@
+using LinearAlgebra
+using WaveOpticsPropagation
+using BenchmarkTools
+using NPZ
+using JSON
+using Printf
+using CUDA  # Comment if you don't have CUDA
+using FFTW
+FFTW.set_num_threads(Sys.CPU_THREADS)
+
+function Base.vec(u::AbstractArray, nd::Integer)
+    @assert nd in (1, 2)
+    @assert ndims(u) >= nd
+    reshape(eachslice(u; dims = Tuple((nd + 1):ndims(u))), :)
+end
+
+path = joinpath(@__DIR__, "../data/test_cases.npz")
+
+data = npzread(path)
+input_modes = data["input_modes"]
+output_modes = data["output_modes"]
+n_modes_orig = size(input_modes, 3)
+
+λ = data["wavelength"]
+z = data["z"]
+dx = data["dx"]
+dy = data["dy"]
+
+nx = size(input_modes, 1)
+
+# Data augmentation to reach limit of memory usage
+N_rep = 6
+n_modes = n_modes_orig * N_rep
+input_modes = repeat(input_modes, 1, 1, N_rep)
+output_modes = repeat(output_modes, 1, 1, N_rep)
+
+# Check correctness
+function check_correctness(uin, uout, nx, dx, λ, z)
+    p_fwd = AngularSpectrum(uin[1], z, λ, nx*dx; padding = false, bandlimit = false)
+    p_bwd = AngularSpectrum(uin[1], -z, λ, nx*dx; padding = false, bandlimit = false)
+
+    function correct()
+        uf = [copy(p_fwd(x; crop = false)) for x in uin]
+        ub = [copy(p_bwd(x; crop = false)) for x in uf]
+        (all(isapprox.(abs2.(dot.(uf, uout)), 1; atol = 1e-4)),
+         all(isapprox.(abs2.(dot.(ub, uin)), 1; atol = 1e-4)))
+    end
+    correct
+end
+
+uin = vec(input_modes, 2)
+uout = vec(output_modes, 2)
+
+fwd_correct, bwd_correct = check_correctness(uin, uout, nx, dx, λ, z)()
+@assert fwd_correct
+@assert bwd_correct
+
+println("✅ WaveOpticsPropagation AS correct!")
+
+function bench_propagate!(uin, nx, dx, λ, z)
+    p_fwd = AngularSpectrum(uin[1], z, λ, nx*dx; padding = false, bandlimit = false)
+    p_bwd = AngularSpectrum(uin[1], -z, λ, nx*dx; padding = false, bandlimit = false)
+
+    function prop()
+        for x in uin
+            uf = p_fwd(x; crop = false)
+            p_bwd(uf; crop = false)
+        end
+    end
+    prop
+end
+
+function bench_propagate(uin, nx, dx, λ, z)
+    bench_propagate!(uin, nx, dx, λ, z)
+end
+
+n_bench = 20
+nmodes_list = Int.(round.(range(1, n_modes, length = n_bench)))
+
+bench_cpu_results = []
+bench_gpu_results = []
+
+println("Benchmarking $(length(nmodes_list)) configurations...")
+println("="^60)
+
+for (i, n_modes) in enumerate(nmodes_list)
+    print("[$i/$(length(nmodes_list))] $n_modes modes: ")
+    GC.gc()
+    CUDA.reclaim()
+
+    # CPU
+    print("CPU...")
+    local bench_cpu = bench_propagate(@view(uin[1:n_modes]), nx, dx, λ, z)
+    b = @benchmark $bench_cpu() seconds=10
+    cpu_time = median(b.times) / 1e6
[…]
