+++ benchmarks/angular_spectrum/Project.toml
@@ -1,3 +1,9 @@
 [deps]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FluxOptics = "5d1b044d-dcff-404c-be27-e1944ce84415"
+JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+WaveOpticsPropagation = "c4c7a1f9-3adc-4a73-843a-d378b6c86436"

[…]
