### 2026-02-16 2db0a7d
Benchmarking JaxOptics and TorchOptics
