+++ benchmarks/angular_spectrum/scripts/bench_fluxoptics_pure.jl
@@ -6,7 +6,9 @@ include("fluxoptics_common.jl")
 println("✅ FluxOptics pure AS correct!")
 
 n_bench = 20
-nmodes_list = Int.(round.(range(1, n_modes, length = n_bench)))
+# nmodes_list = Int.(round.(range(1, n_modes, length = n_bench)))
+nmodes_list = unique([1;
+                      Int.(round.(exp.(range(log(10), log(n_modes), length = n_bench))))])
 
 bench_cpu_results = []
 bench_gpu_results = []
@@ -16,8 +18,6 @@ println("="^60)
 
 for (i, n_modes) in enumerate(nmodes_list)
     print("[$i/$(length(nmodes_list))] $n_modes modes: ")
-    GC.gc()
-    CUDA.reclaim()
 
     # CPU
     print("CPU...")
