+++ benchmarks/angular_spectrum/scripts/plot_results.jl
@@ -1,0 +1,255 @@
+using JSON, CairoMakie, Statistics
+
+results_dir = joinpath(@__DIR__, "../results")
+
+# Load data
+data_fo = JSON.parsefile(joinpath(results_dir, "fluxoptics_bench.json"))
+data_pure = JSON.parsefile(joinpath(results_dir, "fluxoptics_pure_bench.json"))
+data_wave = JSON.parsefile(joinpath(results_dir, "waveopticspropagation_bench.json"))
+data_jax = JSON.parsefile(joinpath(results_dir, "jaxoptics_bench.json"))
+data_torch = JSON.parsefile(joinpath(results_dir, "torchoptics_bench.json"))
+
+# Extraction helper
+function extract(data, device)
+    d = data[device]
+    nmodes = [x["n_modes"] for x in d]
+    times = [x["time_median_ms"] for x in d]
+    return nmodes, times
+end
+
+function interpolate_loglog(nm_ref, t_ref, nm_target)
+    log_nm_ref = log10.(nm_ref)
+    log_t_ref = log10.(t_ref)
+    log_nm_target = log10.(nm_target)
+
+    t_interp = similar(nm_target, Float64)
+
+    for (i, log_nm) in enumerate(log_nm_target)
+        if log_nm <= log_nm_ref[1]
+            t_interp[i] = 10^log_t_ref[1]
+        elseif log_nm >= log_nm_ref[end]
+            t_interp[i] = 10^log_t_ref[end]
+        else
+            j = searchsortedfirst(log_nm_ref, log_nm)
+            if j == 1
+                j = 2
+            end
+
+            x0, x1 = log_nm_ref[j-1], log_nm_ref[j]
+            y0, y1 = log_t_ref[j-1], log_t_ref[j]
+
+            log_t = y0 + (y1 - y0) * (log_nm - x0) / (x1 - x0)
+            t_interp[i] = 10^log_t
+        end
+    end
+
+    return t_interp
+end
+
+function compute_speedup_with_interp(nm_ref, t_ref, nm_test, t_test)
+    t_ref_interp = interpolate_loglog(nm_ref, t_ref, nm_test)
+    speedup = t_ref_interp ./ t_test
+    return nm_test, speedup
+end
+
+# Extract data
+nm_fo, t_fo = extract(data_fo, "gpu")
+nm_pure, t_pure = extract(data_pure, "gpu")
+nm_wave, t_wave = extract(data_wave, "gpu")
+nm_jax, t_jax = extract(data_jax, "gpu")
+nm_torch, t_torch = extract(data_torch, "gpu")
+
+nm_fo_cpu, t_fo_cpu = extract(data_fo, "cpu")
+nm_pure_cpu, t_pure_cpu = extract(data_pure, "cpu")
+nm_wave_cpu, t_wave_cpu = extract(data_wave, "cpu")
+nm_jax_cpu, t_jax_cpu = extract(data_jax, "cpu")
+nm_torch_cpu, t_torch_cpu = extract(data_torch, "cpu")
+
+# Compute speedups vs FluxOptics custom (baseline)
+nm_speedup_pure, speedup_pure = compute_speedup_with_interp(nm_fo, t_fo, nm_pure, t_pure)
+nm_speedup_wave, speedup_wave = compute_speedup_with_interp(nm_fo, t_fo, nm_wave, t_wave)
+nm_speedup_jax, speedup_jax = compute_speedup_with_interp(nm_fo, t_fo, nm_jax, t_jax)
+nm_speedup_torch,
[…]
