+++ benchmarks/angular_spectrum/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FluxOptics = "5d1b044d-dcff-404c-be27-e1944ce84415"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
