### 2026-02-17 e0b5468
Add performance benchmarks section to README.md
Added performance benchmarks comparing FluxOptics.jl with other libraries.
