### 2026-02-17 803d8bc
Update performance notes in README.md
Clarify performance comparison details between FluxOptics, TorchOptics, and JaxOptics.
