+++ README.md
@@ -145,9 +145,12 @@ on Angular Spectrum propagation (512×512, up to 1386 modes, NVIDIA RTX 4070):
 FluxOptics also handles **~1.5× more modes** than JaxOptics before running out of GPU memory.
 
 > **Note**: these benchmarks measure Angular Spectrum propagation performance in isolation.
-> On full optimization loops, the gap may narrow: for instance TorchOptics showed as little as 13×
-> slowdown on a 3-plane monomode conversion task, and XLA-based frameworks like JAX may 
-> even slightly outperform Julia due to whole-graph compilation.
+> On full optimization loops, the gap may narrow: TorchOptics showed as little as 13×
+> slowdown on a 3-plane monomode conversion task. For FFT and tensor-dominated problems,
+> XLA-based frameworks like JAX may even slightly outperform Julia due to whole-graph
+> compilation, and for such cases, JaxOptics may well be sufficient.
+> FluxOptics is expected to show stronger advantages on more complex optical components
+> where manual optimization and Julia's flexibility become decisive.
 
 → [Full benchmark details](benchmarks/README.md)
 
