### 2026-02-17 ca00cf0
Fix link formatting in README.md
