### 2026-02-17 #9734 @anscoil
@mdmaas Following this discussion, I developed [JaxOptics](https://github.com/anscoil/jaxoptics), a minimal JAX/Python reimplementation of FluxOptics core functionality, specifically to establish a fair and well-optimized Python baseline for comparison. This led to comprehensive benchmarks now documented in the [README Performance section](https://github.com/anscoil/FluxOptics.jl#performance), which I believe directly addresses your point 2 by situating FluxOptics within both the Julia and Python ecosystems.

The comparison includes WaveOpticsPropagation.jl, which is to my knowledge the only other actively maintained differentiable optics package in Julia. However, its scope is limited to propagation functions, and it does not support end-to-end system optimization. In that sense, [TorchOptics](https://github.com/MatthewFilipovich/torchoptics) is actually the closest conceptual equivalent to FluxOptics in spirit (full system abstraction, trainable components, optimization loop), just in a different ecosystem and with significantly different performance characteristics.

| Library | GPU speedup vs FluxOptics |
|---------|--------------------------|
| **FluxOptics** | **1×** (baseline) |
| **JaxOptics** | 0.8× |
| **WaveOpticsPropagation.jl** | 0.6× |
| **TorchOptics** | 0.002× |

FluxOptics also handles **~1.5× more modes** than JaxOptics before running out of GPU memory.

> **Note**: these benchmarks measure Angular Spectrum propagation performance in isolation.
> On full optimization loops, the gap may narrow: TorchOptics showed as little as 13×
> slowdown on a 3-plane monomode conversion task. For FFT and tensor-dominated problems,
> XLA-based frameworks like JAX may even slightly outperform Julia due to whole-graph
> compilation, and for such cases, JaxOptics may well be sufficient.
> FluxOptics is expected to show stronger advantages on more complex optical components
> where manual optimization and Julia's flexibility become decisive.

→ [Full benchmark details](https://github.com/anscoil/FluxOptics.jl/tree/main/benchmarks)
