### 2026-02-17 #9734 @abhishektiwari
Thanks @anscoil

I want to confirm that [JaxOptics](https://github.com/anscoil/jaxoptics) repository is not in scope of this review. If so I would like to move ahead to archiving tasks given both reviewers have already given thumbs up.
