### 2026-02-17 #9734 @anscoil
@abhishektiwari No JaxOptics is not in scope of this review, it is a quick separate project I did to make a fair benchmark comparison with the Python ecosystem and JAX which is considered as a standard but I did not mention in the JOSS paper.
