### 2026-02-17 0ba4e7d
Bump version from 0.1.0 to 0.1.2
