### 2026-02-17 5d8e52c
Update Mode Sorting link to stable version
