### 2026-02-17 #9734 @anscoil
@abhishektiwari All archival tasks completed:
* [x]  Double check Paper authors and affiliations (including ORCIDs)
* [x]  Make a tagged release of your software on Github, and list the version tag of the archived version here.
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct title (should match the paper title).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct authors and affiliations (should match the paper authors including order).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct version (should match the above Github tag you created).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct ORCIDs (should match the paper authors).
* [x]  Please list the DOI of the archived version here: **v0.1.2**: https://doi.org/10.5281/zenodo.18673900
