### 2026-02-17 e92adfa
Fix stable symlink to point to v0.1.2
