+++ v1.0.2/siteinfo.js
@@ -1,2 +1,0 @@
-var DOCUMENTER_CURRENT_VERSION = "v1.0.2";
-var DOCUMENTER_IS_DEV_VERSION = false;
