### 2026-02-18 #9734 @abhishektiwari
@anscoil  Please see minor edits for paper,

Main Body (Edits in paper.md):

Line 7–8: "which consists in determining the structure". Consider "which involves determining the structure"
Line 26: "at computational cost comparable" should be "at a computational cost."
Line 92–93: "Comparison with TorchOptics… on a beam splitter inverse design task". Consider: "We compared FluxOptics.jl with TorchOptics on a beam splitter inverse design task…"

Also, following section seems redundant so consider removing.

```
Aknowledgments

There are no acknowledgments to declare and no conflicts of interest to disclose.
```

References (Edits in paper.bib): Please use {} explicitly to capitalise as by default JOSS formats reference titles as sentence case in PDF rendering.

Line 137: Filipovich & Lvovsky (2024): "python" should be Python; "fourier" should be "Fourier"
Line 148: Hughes et al. (2019): "maxwell's" should be "Maxwell's"
Line 139: Fontaine et al. (2019): "Laguerre-gaussian" should be "Laguerre-Gaussian"
