### 2026-02-18 673fffe
JOSS minor corrections
