+++ paper/paper.bib
@@ -23,7 +23,7 @@
 }
 
 @article{Hughes2018,
-  title={Forward-mode differentiation of Maxwell's equations},
+  title={Forward-mode differentiation of {M}axwell's equations},
   author={Hughes, Tyler W and Williamson, Ian AD and Minkov, Momchil and Fan, Shanhui},
   journal={ACS Photonics},
   volume={6},
@@ -75,7 +75,7 @@ doi = {10.1364/OE.521322},
 }
 
 @misc{TorchOptics,
-      title={TorchOptics: An open-source Python library for differentiable Fourier optics simulations},
+      title={TorchOptics: An open-source {P}ython library for differentiable {F}ourier optics simulations},
       author={Matthew J. Filipovich and A. I. Lvovsky},
       year={2024},
       eprint={2411.18591},
@@ -220,7 +220,7 @@ doi = {10.1364/OE.521322},
 }
 
 @article{Fontaine2019,
-  title={Laguerre-Gaussian mode sorter},
+  title={Laguerre-{G}aussian mode sorter},
   author={Fontaine, Nicolas K and Ryf, Roland and Chen, Haoshuo and Neilson, David T and Kim, Kwangwoong and Carpenter, Joel},
   journal={Nature communications},
   volume={10},
