+++ versions.js
@@ -1,8 +1,7 @@
 var DOC_VERSIONS = [
   "stable",
-  "v1.0",
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.0.2";
+var DOCUMENTER_NEWEST = "v0.1.2";
 var DOCUMENTER_STABLE = "stable";
