### 2026-02-18 #9734 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
