### 2026-02-23 95625a7
minor changes in paper.md
