### 2026-02-23 #9734 @danielskatz
👋 @anscoil - I've suggested some minor changes in https://github.com/anscoil/FluxOptics.jl/pull/5 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication
