### 2026-02-23 #9734 @danielskatz
@anscoil - sorry, I added a typo - fixed in https://github.com/anscoil/FluxOptics.jl/pull/6
