### 2026-02-23 #9734 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
