### 2026-02-23 1e2b354
Simplification for better output (4 pages)
