### 2026-02-23 #9734 @anscoil
@danielskatz I made a minor modification after the las PR merge to avoid using a fifth page just for a DOI, I don't know if this was taken into account... It is just for a nicer output.
