### 2026-02-23 #9734 @danielskatz
Congratulations to @anscoil (Nicolas BARRE) on your JOSS publication!!

JOSS completely depends on volunteers - we couldn't do this without you.
